+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - AppStore app: don't update UI after the user has closed the app
 - Wifi app: cleanup and improve keyboard and focus handling
 - Wifi app: improve different screensize handling
+- About app: enable scrolling using arrow keys so off-screen info can be viewed
 
 0.1.1
 =====